bio_mcp/requirements.txt
# Bio MCP Requirements
mcp>=1.0.0
propka
biopython
fastapi
uvicorn
websockets
aiofiles
anthropic
openai
google-generativeai
dashscope
pydantic
pytest
pytest-asyncio
python-dotenv